Fix getColumnType
First of all, this is way simpler (the TableProxy interface only allows
you to get the data type as a string, as opposed to as an enumerated
value). However, there seems to be a problem with the way I was passing
the C strings to Julia. It seems that the strings were occasionally
being freed and overwritten. I do not yet have a solution for this
problem.