minimal units implementation
SIUnits is so close to being awesome, but it always seems to fall short
of my expectations when I try to use it.

Instead I implement an internal units implementation. The user will
occasionally have to specify the units associated with the values they
are passing to measures, but should never have to deal with my
hand-rolled units implementation because it's never part of the return
value.