switch from my crappy units implementation to Unitful
I'm still not entirely happy with this because Unitful encourages people
to edit their unit definitions at the package level, which means
`using Unitful` can have different results depending on the machine.
Therefore you can't even reliably expect that there are units for a day,
a second, or any time units at all.

This is not really acceptable so this needs to change before I merge
this into master.

I've brought up the issue at
https://github.com/ajkeller34/Unitful.jl/issues/33

The likely outcome of this is that I will maintain an AstroUnits package
that will extend Unitful and provide a good set of units that are
suitable for general use in astronomy. CasaCore will then depend on that
package instead.